Divide structs into files
