+++ src/tf_error.rs
@@ -1,0 +1,12 @@
+/// Enumerates the different types of errors
+#[derive(Clone, Debug)]
+pub enum TfError {
+    /// Error due to looking up too far in the past. I.E the information is no longer available in the TF Cache.
+    AttemptedLookupInPast,
+    /// Error due ti the transform not yet being available.
+    AttemptedLookUpInFuture,
+    /// There is no path between the from and to frame.
+    CouldNotFindTransform,
+    /// In the event that a write is simultaneously happening with a read of the same tf buffer
+    CouldNotAcquireLock,
+}
