+++ src/transforms.rs
@@ -213,3 +213,20 @@ mod test {
         assert_eq!(interpolate(tf1, tf2, 0.5), expected);
     }
 }
+
+pub(crate) fn to_transform_stamped(
+    tf: Transform,
+    from: std::string::String,
+    to: std::string::String,
+    time: rosrust::Time,
+) -> TransformStamped {
+    TransformStamped {
+        header: Header {
+            frame_id: from,
+            stamp: time,
+            seq: 1u32,
+        },
+        child_frame_id: to,
+        transform: tf,
+    }
+}
